refactor: separate picking a photo and uploading a photo into two classes.
